major update to Bot 1.0.1.1 defined under Milestone under the same name. Fixes quotes/macros, updates docs, and adds some new mod commands

+++ src/io/ph/bot/commands/general/Quote.java
@@ -7,6 +7,10 @@ import org.apache.commons.lang3.StringUtils;
 import java.util.List;
 import java.time.LocalDate;
 import java.lang.*;
+import java.util.Random;
+import java.util.ArrayList;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.lang.Integer;
 
 import io.ph.bot.commands.Command;
 import io.ph.bot.commands.CommandCategory;
@@ -19,8 +23,10 @@ import io.ph.util.Util;
 import io.ph.bot.Bot;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.entities.Member;
-import net.dv8tion.jda.core.entities.Message;
 import net.dv8tion.jda.core.entities.Guild;
+import net.dv8tion.jda.core.entities.Message;
+import net.dv8tion.jda.core.entities.MessageHistory.MessageRetrieveAction;
+import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
 /**
  * Create, delete, search, call information, and call Quotes
@@ -34,21 +40,25 @@ import net.dv8tion.jda.core.entities.Guild;
         category = CommandCategory.FUN,
         permission = Permission.NONE,
         description = "Create, delete, edit, search, or get information on a quote\n"
-                + "Quotes are text from user that are saved to a db to be called later",
-        example = "create \"QuotedUser\" contents *This creates a quote for user `QuotedUser`*\n"
+                + "Quotes are text from user that are saved to a db to be called later\n"
+                + "If user doesn't exist in server, set the quote to the bot\n",
+        example = "------ *If no arg will select random quote*\n"
+                + "previous/p *Makes a quote of the previous message*\n"
+                + "user/u Username *Pulls random quote from user*\n"
+                + "create \"QuotedUser\" contents *This creates a quote for user `QuotedUser`*\n"
                 + "delete 57 *This deletes the quote 57*\n"
-                + "edit \"57\" new contents *This edits the quote #57's contents*\n"
-                + "list \"QuotedUser\" *lists all quotes created by the QuotedUser*\n"
-                + "showall *lists all quotes for server*\n"
+                + "edit \"57\" username new contents *This edits the quote #57's contents and user*\n"
+                + "list \"QuotedUser\" *Lists all quotes created by the QuotedUser*\n"
+                + "showall *Lists all quotes for server*\n"
                 + "info 57 *This gives information on the quote #57*\n"
-                + "top *Will rank the top quofte*\n"
+                + "top *Ranks the top quote*\n"
                 //+ "rank *Will rank the top 10 quotes*\n"
-                + "if no arg will select random quote \n"
         )
 public class Quote extends Command {
     private EmbedBuilder em;
     private Message msg;
     private String contents;
+    private String param;
     private int uniq;
 
     @Override
@@ -56,17 +66,22 @@ public class Quote extends Command {
         this.msg = msg;
         this.em = new EmbedBuilder();
         this.contents = Util.getCommandContents(msg);
-        String param = "";
+        this.param = "";
         
         try{
             param = Util.getParam(msg);
         } catch (IndexOutOfBoundsException e) {
             param = "";
         }   
+        //debug
+        //System.out.println("Contents: " + contents);
+        //System.out.println("Param: " + param);
         if(param.equalsIgnoreCase("create")) {
             createQuote();
         } else if(param.equalsIgnoreCase("delete")) {
             deleteQuote();
+        } else if(param.equalsIgnoreCase("user") || param.equalsIgnoreCase("u")) {
+            userQuote();
         } else if(param.equalsIgnoreCase("edit")) {
             editQuote();
         } else if(param.equalsIgnoreCase("list")) {
@@ -75,31 +90,29 @@ public class Quote extends Command {
             listQuotesAll();
         } else if (param.equalsIgnoreCase("info")) {
             quoteInfo();
+        } else if (param.equalsIgnoreCase("previous") || param.equalsIgnoreCase("p")) {
+            previousMsg();
         } else if(param.equalsIgnoreCase("top")) {
+            quoteTop();
+        } else if(param.equalsIgnoreCase("rank")) {
             quoteRank();
-        //} else if(param.equalsIgnoreCase("rank")) {
-        //   quoteRank();
         } else if(!contents.equals("")){
                 QuoteObject m = QuoteObject.forName(Integer.parseInt(contents), msg.getGuild().getId().toString(), true);
                 msg.getChannel().sendMessage("#" + Integer.toString(m.getQuoteUniq()) 
                     + " \"" + m.getQuoteContent() + "\" ~ " 
-                    + m.getFallbackUsername().getUser().getName() 
-                    + "#" 
-                    + m.getFallbackUsername().getUser().getDiscriminator() 
+                    + Util.resolveNameFromMember(m.getFallbackUsername(),false)
                     + " on " 
                     + m.getDate().toString()).queue(success -> {msg.delete().queue();});
                 return;
         } else {
             // send random quote
             try {
-                String[] m = QuoteObject.chooseRandomQuote(msg.getGuild().getId());
-                msg.getChannel().sendMessage("#" + m[0] 
-                    + " \"" + m[3] + "\" ~ " 
-                    + Util.resolveUserFromMessage(m[2], msg.getGuild().getId()).getUser().getName() 
-                    + "#" 
-                    + Util.resolveUserFromMessage(m[2], msg.getGuild().getId()).getUser().getDiscriminator() 
+                QuoteObject m = QuoteObject.chooseRandomQuote(msg.getGuild().getId().toString());
+                msg.getChannel().sendMessage("#" + Integer.toString(m.getQuoteUniq()) 
+                    + " \"" + m.getQuoteContent() + "\" ~ " 
+                    + Util.resolveNameFromMember(m.getFallbackUsername(),false)
                     + " on " 
-                    + LocalDate.parse(m[1]).toString()).queue(success-> {msg.delete().queue();});
+                    + m.getDate().toString()).queue(success -> {msg.delete().queue();});
                 return;
             } catch (IllegalArgumentException e) {
                 em.setTitle("Error", null)
@@ -109,7 +122,50 @@ public class Quote extends Command {
         }
         // debug
         //System.out.println(param);
-        msg.getChannel().sendMessage(em.build()).queue(success -> {msg.delete().queue();});
+        if (!em.isEmpty()){
+            msg.getChannel().sendMessage(em.build()).queue(success -> {msg.delete().queue();});
+        }
+        msg.delete().queue();
+    }
+
+    /**
+     * Create a quote from a previous message
+     */
+    private void previousMsg() {
+        Message nMsg = msg.getChannel().getHistoryBefore(msg, 1).complete().getRetrievedHistory().get(0);
+        // System.out.println(nMsg);
+        String nContents = nMsg.getContentDisplay();
+        String nAuthor = Util.combineStringArray(
+                         Util.removeLastArrayEntry(
+                         Util.resolveNameFromMember(
+                         Util.memberFromMessage(nMsg),true).split("#")));
+        String tContents ="\"" + nAuthor + "\" " + nContents;
+        //debug
+        //System.out.println("nAuthor: "+nAuthor);
+        System.out.println("tContents:"+tContents);
+        createQuote(tContents);
+    }
+    /**
+     * Grab a random quote from the user
+     */
+    private void userQuote() {
+
+        Random randomGenerator = new Random();
+        contents = Util.getCommandContents(contents);
+        String mem = Util.resolveMemberFromMessage(contents,msg.getGuild()).getUser().getId();
+        String[] userUniqs = QuoteObject.searchByUser(mem,msg.getGuild().getId());
+        int size = userUniqs.length;
+        int uniq = Integer.parseInt(userUniqs[randomGenerator.nextInt(size)]);
+        //debug
+        //System.out.println(size);
+        //System.out.println(uniq);
+        QuoteObject m = QuoteObject.forName(uniq,msg.getGuild().getId(),true);
+        msg.getChannel().sendMessage("#" + Integer.toString(m.getQuoteUniq()) 
+            + " \"" + m.getQuoteContent() + "\" ~ " 
+            + Util.resolveNameFromMember(m.getFallbackUsername(),false)
+            + " on " 
+            + m.getDate().toString()).queue(success -> {msg.delete().queue();});
+        return;
     }
 
     /**
@@ -117,6 +173,13 @@ public class Quote extends Command {
      */
     private void createQuote() {
         contents = Util.getCommandContents(contents);
+        createQuote(contents);
+    }
+
+    /**
+     * Create a Quote Object
+     */
+    private void createQuote(String contents) {
         //contents = "Nick testing the create quote commands";
         if(contents.equals("") || contents.split(" ").length < 2) {
             em.setTitle("Error", null)
@@ -171,7 +234,7 @@ public class Quote extends Command {
             if(m.delete(msg.getAuthor().getId())) {
                 em.setTitle("Success", null)
                 .setColor(Util.resolveColor(Util.memberFromMessage(msg), Color.GREEN))
-                .setDescription("Quote ** #$" + contents + "** deleted");
+                .setDescription("Quote ** #" + contents + "** deleted");
             } else {
                 em.setTitle("Error", null)
                 .setColor(Color.RED)
@@ -209,9 +272,9 @@ public class Quote extends Command {
             String userID = Util.resolveMemberFromMessage(resolved[1].split(" ")[0],msg.getGuild()).getUser().getId();
             String newContent = Util.getCommandContents(resolved[1]);
             // debug
-            System.out.println(uniq);
-            System.out.println(userID);
-            System.out.println(newContent);
+            //System.out.println(uniq);
+            //System.out.println(userID);
+            //System.out.println(newContent);
 
             QuoteObject m = QuoteObject.forName(uniq, msg.getGuild().getId());
             if(m.edit(msg.getAuthor().getId(), newContent,userID)) {
@@ -235,7 +298,7 @@ public class Quote extends Command {
      * List all quotes made by a user
      */
     private void listQuotes() {
-        Member m;
+        Member m ;
         String possibleUser = Util.getCommandContents(this.contents);
         if (possibleUser.isEmpty()) {
             m = msg.getMember();
@@ -243,10 +306,12 @@ public class Quote extends Command {
             m = Util.resolveMemberFromMessage(possibleUser, msg.getGuild());
         }
         String[] results = QuoteObject.searchByUser(m.getUser().getId(), msg.getGuild().getId());
+        //System.out.println("User: " + m);
+        //System.out.println("Cont: " + results);
         if (results == null) {
             em.setTitle("Error", null)
             .setColor(Color.RED)
-            .setDescription("No quotes found for user " + m.getEffectiveName());
+            .setDescription("No quotes found for user " + Util.resolveNameFromMember(m,false));
             return;
         }
         StringBuilder sb = new StringBuilder();
@@ -256,32 +321,47 @@ public class Quote extends Command {
         
         String finalSb = sb.toString().substring(0, sb.toString().length() - 2);
         
-        em.setTitle("Quotes created by " + m.getEffectiveName(), null)
+        em.setTitle("Quotes created by " + Util.resolveNameFromMember(m,false), null)
         .setColor(Util.resolveColor(Util.memberFromMessage(msg), Color.GREEN))
         .setDescription(finalSb);
     }
 
     /**
-     * List all quotes made
+     * List all quotes made by either the guild or by a user
      */
     private void listQuotesAll() {
-        StringBuilder totalString = new StringBuilder();
-        List<Member> iteratorion = msg.getGuild().getMembers();
-        iteratorion.forEach(j -> {
-            String[] results = QuoteObject.searchByUser(j.getUser().getId(),msg.getGuild().getId());
+        String nContent = Util.getCommandContents(Util.getCommandContents(msg));
+        ArrayList<String> totalString = new ArrayList<String>(1);
+        //debug
+        //System.out.println("Cont: " + nContent);
+        if (!nContent.equals("")) {
+            Member qMember = Util.resolveMemberFromMessage(nContent,msg.getGuild());
+            String name = Util.resolveNameFromMember(qMember,false);
+            em.setTitle("All Quotes created by: **" + name + "**", null);
+            String[] results = QuoteObject.searchByUser(qMember.getUser().getId(),msg.getGuild().getId());
             if (results != null) {
-                StringBuilder sb = new StringBuilder();
                 for (String s : results) {
-                    sb.append(s + ", ");
+                    QuoteObject qO = QuoteObject.forName(Integer.parseInt(s),msg.getGuild().getId());
+                    totalString.add("**#" + qO.getQuoteUniq() + "** \"" + qO.getQuoteContent() + "\"");
                 }
-                String fin = sb.toString();
-                totalString.append(j.getUser().getName() + ": " + fin.substring(0, fin.length() - 2) + "\n");
             }
-        });
-        String finalTotal = totalString.toString();
-        em.setTitle("All Quotes created in this guild: ", null)
-        .setColor(Util.resolveColor(Util.memberFromMessage(msg), Color.GREEN))
-        .setDescription(finalTotal);
+        } else {
+            em.setTitle("All Quotes created in this guild: ", null);
+            List<Member> iteratorion = msg.getGuild().getMembers();
+            iteratorion.forEach(j -> {
+                String[] results = QuoteObject.searchByUser(j.getUser().getId(),msg.getGuild().getId());
+                if (results != null) {
+                    StringBuilder sb = new StringBuilder();
+                    for (String s : results) {
+                        sb.append(s + ", ");
+                    }
+                    String fin = sb.toString();
+                    totalString.add("**" + Util.resolveNameFromMember(j,false) + ":** " + fin.substring(0, fin.length() - 2));
+                }
+            });
+        }
+        MessageUtils.staggerArray(totalString,msg,em);
+        em.clear();
     }
 
     /**
@@ -302,10 +382,9 @@ public class Quote extends Command {
         try {
             int uniq = Integer.parseInt(contents);
             QuoteObject m = QuoteObject.forName(uniq, msg.getGuild().getId());
-            Member mem = msg.getGuild().getMemberById(m.getUserID());
             em.setTitle("Information on #" + contents, null)
             .setColor(Util.resolveColor(Util.memberFromMessage(msg), Color.GREEN))
-            .addField("Creator", m.getFallbackUsername().getUser().getName() + "#" + m.getFallbackUsername().getUser().getDiscriminator(), true)
+            .addField("Creator", Util.resolveNameFromMember(m.getFallbackUsername(),false), true)
             .addField("Content", m.getQuoteContent(), true)
             .addField("Hits", m.getHits()+"", true)
             .addField("Date created", m.getDate().toString(), true);
@@ -347,12 +426,12 @@ public class Quote extends Command {
     /**
      * Send information on a macro
      */
-    private void quoteRank() {
+    private void quoteTop() {
         try {
             QuoteObject m = QuoteObject.topQuote(msg.getGuild().getId());
             em.setTitle("Information on #" + Integer. toString(m.getQuoteUniq()), null)
             .setColor(Util.resolveColor(Util.memberFromMessage(msg), Color.GREEN))
-            .addField("Creator", m.getFallbackUsername().getUser().getName() + "#" + m.getFallbackUsername().getUser().getDiscriminator(), true)
+            .addField("Creator", Util.resolveNameFromMember(m.getFallbackUsername(),false), true)
             .addField("Content", m.getQuoteContent()+"", true)
             .addField("Hits", m.getHits()+"", true)
             .addField("Date created", m.getDate().toString(), true);
@@ -362,21 +441,49 @@ public class Quote extends Command {
             .setDescription(e.getMessage());
         }
     }
+
+    /**
+     * Send information on a quote
+     */
+    private void quoteRank() {
+        try {
+            ArrayList<String> rQ = QuoteObject.rankQuote(msg.getGuild().getId());
+            ArrayList<String> message = new ArrayList<String>(10);
+            // debug
+            //System.out.println("rQ: " + rQ);
+            for (int i = 0; i < rQ.size(); i++) {
+                QuoteObject m = QuoteObject.forName(Integer.parseInt(rQ.get(i)),msg.getGuild().getId());
+                message.add("#" + Integer.toString(m.getQuoteUniq()) 
+                    + " \"" + m.getQuoteContent() + "\" ~ " 
+                    + Util.resolveNameFromMember(m.getFallbackUsername(),false)
+                    + " on " 
+                    + m.getDate().toString()
+                    + "\n");
+            }
+            MessageUtils.sendMessage(msg.getChannel().getId(),Util.combineStringArray(message));
+        } catch (IllegalArgumentException e) {
+            em.setTitle("Error", null)
+            .setColor(Color.RED)
+            .setDescription(e.getMessage());
+        }
+    }
+
     /**
      * Resolve Quote Username and contents from a create statement
      * This works to involve quotations around a spaced quote name
      * @param s The parameters of a create statement - The contents past the $quote create bit
-     * @return Multi index array: [0] userid, [1] is the contents
+     * @return Multi index array: [0] username, [1] is the contents
      * Prerequisite: s.split() must have length of >= 2
      */
     private static String[] resolveQuoteUserAndContents(String s) {
         // debug
         //System.out.println(s);
         String[] toReturn = new String[2];
-        if(s.contains("\"") && StringUtils.countMatches(s, "\"") > 1) {
-            int secondIndexOfQuotes = s.indexOf("\"", s.indexOf("\"") + 1);
-            toReturn[0] = s.substring(s.indexOf("\"") + 1, secondIndexOfQuotes);
-            toReturn[1] = s.substring(secondIndexOfQuotes + 2);
+        if(s.contains("\"") && s.split("\"").length > 1) {
+            toReturn[0] = s.split("\"")[1];
+            toReturn[1] = Util.combineStringArray(
+                          Util.removeFirstArrayEntry(
+                            s.split("\"")));
         } else {
             toReturn[0] = s.split(" ")[0];
             toReturn[1] = Util.getCommandContents(s);
